+++ core/dbt/events/types.py
@@ -137,7 +137,7 @@ class MainReportArgs(DebugLevel, Cli, File):
 
 
 @dataclass
-class MainTrackingUserState(DebugLevel, Cli):
+class MainTrackingUserState(DebugLevel, Cli, File):
     user_state: str
     code: str = "A003"
 
@@ -2365,7 +2365,7 @@ class DisableTracking(WarnLevel, Cli, File):
 
 
 @dataclass
-class SendingEvent(DebugLevel, Cli):
+class SendingEvent(DebugLevel, Cli, File):
     kwargs: str
     code: str = "Z040"