+++ core/dbt/task/base.py
@@ -15,6 +15,7 @@ from dbt.exceptions import (
     InternalException
 )
 from dbt.logger import log_manager
+import dbt.events.functions as event_logger
 from dbt.events.functions import fire_event
 from dbt.events.types import (
     DbtProjectError, DbtProjectErrorException, DbtProfileError, DbtProfileErrorException,
@@ -64,6 +65,9 @@ class BaseTask(metaclass=ABCMeta):
         """A hook called before the task is initialized."""
         if args.log_format == 'json':
             log_manager.format_json()
+            # we're mutating the initialized, but not-yet-configured event logger
+            # because it's being configured too late -- bad! TODO refactor!
+            event_logger.format_json = True
         else:
             log_manager.format_text()
 
@@ -71,6 +75,9 @@ class BaseTask(metaclass=ABCMeta):
     def set_log_format(cls):
         if flags.LOG_FORMAT == 'json':
             log_manager.format_json()
+            # we're mutating the initialized, but not-yet-configured event logger
+            # because it's being configured too late -- bad! TODO refactor!
+            event_logger.format_json = True
         else:
             log_manager.format_text()